+++ KiloMart.DataAccess/Database/ProductOfferDb.cs
@@ -11,7 +11,7 @@ namespace KiloMart.DataAccess.Database
             decimal offPercentage,
             DateTime fromDate,
             DateTime? toDate,
-            float quantity,
+            decimal quantity,
             int provider,
             IDbTransaction? transaction = null)
         {
@@ -40,7 +40,7 @@ namespace KiloMart.DataAccess.Database
             decimal offPercentage,
             DateTime fromDate,
             DateTime? toDate,
-            float quantity,
+            decimal quantity,
             int provider,
             bool isActive,
             IDbTransaction? transaction = null)
@@ -116,7 +116,7 @@ namespace KiloMart.DataAccess.Database
         public decimal OffPercentage { get; set; }
         public DateTime FromDate { get; set; }
         public DateTime? ToDate { get; set; }
-        public float Quantity { get; set; }
+        public decimal Quantity { get; set; }
         public int Provider { get; set; }
         public bool IsActive { get; set; }
     }
