+++ KiloMart.Presentation/Controllers/Domains/Customers/Models.cs
@@ -2,11 +2,11 @@ namespace KiloMart.Presentation.Controllers.Domains.Customers;
 public class CartItemRequest
 {
     public int Product { get; set; }
-    public float Quantity { get; set; }
+    public decimal Quantity { get; set; }
 }
 
 public class UpdateCartItemRequest
 {
     public int? Product { get; set; }
-    public float? Quantity { get; set; }
+    public decimal? Quantity { get; set; }
 }
